chore: cleanup and add commit hook
