+++ tests/package.json
@@ -3,7 +3,9 @@
     "./*"
   ],
   "scripts": {
-    "test": "jest"
+    "test": "jest",
+    "prepare": "node prepare",
+    "posttest": "node prepare --post"
   },
   "dependencies": {
     "jest": "latest"
