+++ tests/with-jest--latest/package.json
@@ -7,7 +7,7 @@
   "devDependencies": {
     "@tests/utils": "*",
     "eslint": "latest",
-    "eslint-config-peppy": "file:../eslint-config-peppy-2.0.0-beta.5.tgz",
+    "eslint-config-peppy": "*",
     "jest": "latest"
   }
 }
