+++ tests/with-jest--v28/package.json
@@ -7,7 +7,7 @@
   "devDependencies": {
     "@tests/utils": "*",
     "eslint": "latest",
-    "eslint-config-peppy": "file:../eslint-config-peppy-2.0.0-beta.5.tgz",
+    "eslint-config-peppy": "*",
     "jest": "28.0.0"
   }
 }
