+++ tests/with-next--v12/package.json
@@ -14,6 +14,6 @@
   "devDependencies": {
     "@tests/utils": "*",
     "eslint": "latest",
-    "eslint-config-peppy": "file:../eslint-config-peppy-2.0.0-beta.5.tgz"
+    "eslint-config-peppy": "*"
   }
 }
