+++ tests/with-peppy/package.json
@@ -7,6 +7,6 @@
   "devDependencies": {
     "@tests/utils": "*",
     "eslint": "latest",
-    "eslint-config-peppy": "file:../eslint-config-peppy-2.0.0-beta.5.tgz"
+    "eslint-config-peppy": "*"
   }
 }
