+++ tests/with-react--latest/package.json
@@ -9,6 +9,6 @@
   "devDependencies": {
     "@tests/utils": "*",
     "eslint": "latest",
-    "eslint-config-peppy": "file:../eslint-config-peppy-2.0.0-beta.5.tgz"
+    "eslint-config-peppy": "*"
   }
 }
