--- turbo.json
@@ -1,18 +1,0 @@
-{
-  "$schema": "https://turborepo.org/schema.json",
-  "pipeline": {
-    "build": {
-      "outputs": ["dist/**"]
-    },
-    "lint": {
-      "outputs": [],
-      "cache": false
-    },
-    "dev": {
-      "cache": false
-    },
-    "test": {
-      "cache": false
-    }
-  }
-}
